refactor: Extract Mui theme in different files
