+++ react/MuiCozyTheme/theme/defaultValues.js
@@ -1,0 +1,13 @@
+const defaultValues = {
+  borderRadius: 6,
+  dialog: {
+    sm: {
+      padding: 16
+    },
+    md: {
+      padding: 32
+    }
+  }
+}
+
+export default defaultValues
