+++ react/MuiCozyTheme/theme/inverted/overrides.js
@@ -1,0 +1,98 @@
+import merge from 'lodash/merge'
+
+import { getCssVariableValue } from '../../../utils/color'
+import makeOverrides from '../overrides'
+import { normalTheme } from '../index'
+import invertedPalette from './palette'
+
+const makeInvertedThemeOverrides = theme => ({
+  ...makeOverrides(theme),
+  MuiOutlinedInput: {
+    root: {
+      boxSizing: 'border-box',
+      '&$focused $notchedOutline': {
+        borderColor: theme.palette.text.primary,
+        borderWidth: '0.0625rem'
+      },
+      '& $notchedOutline': {
+        borderColor: theme.palette.text.primary
+      }
+    }
+  },
+  MuiFormLabel: {
+    root: {
+      '&$focused': {
+        color: theme.palette.text.primary
+      }
+    }
+  },
+  MuiTab: {
+    root: {
+      color: 'rgb(255,255,255) !important'
+    },
+    // This overrides the disabled color of the MuiTab
+    textColorPrimary: {
+      color: 'rgb(255,255,255) !important',
+      opacity: 0.64
+    },
+    textColorSecondary: {
+      color: 'rgb(255,255,255)',
+      opacity: 0.64
+    },
+    selected: {
+      opacity: 1
+    }
+  },
+  MuiSwitch: merge({}, normalTheme.overrides.MuiSwitch, {
+    switchBase: {
+      color: getCssVariableValue('primaryContrastTextColor')
+    },
+    colorPrimary: {
+      '& + $track': {
+        backgroundColor: getCssVariableValue('primaryContrastTextColor')
+      },
+
+      '&$checked': {
+        '& + $track': {
+          boxSizing: 'border-box',
+          backgroundColor: '#34CE68'
+        }
+      }
+    }
+  }),
+  MuiLinearProgress: {
+    colorPrimary: {
+      backgroundColor: 'rgba(255,255,255,0.2)'
+    },
+    colorSecondary: {
+      backgroundColor: 'rgba(255,255,255,0.2)'
+    }
+  },
+  MuiListItem: {
+    button: {
+      '&$selected, &$selected:hover': {
+        backgroundColor: invertedPalette.background.selected
+      },
+      '&:hover': {
+        backgroundColor: invertedPalette.background.selected
+      },
+      '&:focus': {
+        backgroundColor: invertedPalette.background.selected
+      }
+    }
+  },
+  MuiCheckbox: {
+    colorPrimary: {
+      '&.Mui-checked:not(.Mui-disabled)': {
+        color: theme.palette.success.main
+      }
+    },
+    colorSecondary: {
+      '&.Mui-checked:not(.Mui-disabled)': {
+        color: theme.palette.error.main
+      }
+    }
+  }
+})
+
+export default makeInvertedThemeOverrides
