+++ react/MuiCozyTheme/theme/overrides.js
@@ -1,0 +1,449 @@
+import { getCssVariableValue } from '../../utils/color'
+import defaultValues from './defaultValues'
+
+const SWITCH_BAR_WIDTH = 25
+
+const makeOverrides = theme => ({
+  MuiOutlinedInput: {
+    root: {
+      '&$disabled': {
+        background: theme.palette.grey[100]
+      },
+      '&$focused $notchedOutline': {
+        borderWidth: '0.0625rem'
+      },
+      '&:hover $notchedOutline': {
+        borderColor: theme.palette.grey[300]
+      }
+    },
+    notchedOutline: {
+      borderColor: theme.palette.grey[200]
+    }
+  },
+  MuiButton: {
+    root: {
+      height: '2.5rem',
+      padding: '0px 1rem'
+    },
+    sizeSmall: {
+      height: '2rem',
+      padding: '0px 0.75rem'
+    },
+    sizeLarge: {
+      padding: '0px 1.25rem',
+      height: '3rem'
+    },
+    outlined: {
+      borderRadius: 2,
+      minWidth: 112
+    },
+    contained: {
+      borderRadius: 2,
+      boxShadow: 0,
+      minWidth: 112
+    },
+    containedSecondary: {
+      color: 'white'
+    }
+  },
+  MuiTab: {
+    // This overrides the disabled color of the MuiTab
+    textColorPrimary: {
+      color: theme.palette.grey[300]
+    },
+    textColorSecondary: {
+      color: theme.palette.grey[300]
+    },
+    root: {
+      fontSize: theme.typography.subtitle1.fontSize,
+      [theme.breakpoints.up('md')]: {
+        fontSize: theme.typography.subtitle1.fontSize
+      }
+    }
+  },
+  MuiAccordion: {
+    rounded: {
+      borderRadius: defaultValues.borderRadius
+    },
+    root: {
+      boxShadow: '0 4px 12px 0 rgba(0, 0, 0, 0.08)',
+      borderWidth: '0.0625rem',
+      borderStyle: 'solid',
+      borderColor: theme.palette.grey[200],
+      overflow: 'hidden',
+      marginBottom: '1rem'
+    }
+  },
+  MuiAccordionSummary: {
+    expanded: {},
+    root: {
+      backgroundColor: theme.palette.grey[100],
+      textTransform: 'uppercase',
+      fontWeight: 'bold',
+      fontSize: '0.875rem',
+      minHeight: '3.5rem',
+      padding: 0,
+      color: theme.palette.text.primary,
+      '&$expanded': {
+        minHeight: '3.5rem'
+      }
+    },
+    expandIcon: {
+      order: 0,
+      '&&': {
+        marginLeft: '0.3125rem'
+      },
+      transform: 'rotate(-90deg)',
+      '&$expanded': {
+        marginLeft: '0.3125rem',
+        transform: 'rotate(0)'
+      }
+    },
+    content: {
+      margin: '0.75rem 0',
+      paddingLeft: '0.5rem',
+      paddingRight: '0.25rem',
+      order: 1,
+      '& > :last-child': {
+        paddingRight: 0
+      },
+      '&$expanded': {
+        margin: '0.75rem 0'
+      }
+    }
+  },
+  MuiAccordionDetails: {
+    root: {
+      padding: 0,
+      borderTop: `0.0625rem solid ${theme.palette.grey[200]}`
+    }
+  },
+  MuiStepConnector: {
+    line: {
+      borderColor: theme.palette.grey[300]
+    }
+  },
+  MuiList: {
+    padding: {
+      paddingTop: 0,
+      paddingBottom: 0
+    }
+  },
+  MuiListItemIcon: {
+    root: {
+      minWidth: 'auto',
+      marginRight: '1rem',
+      padding: 0,
+      width: 32,
+      justifyContent: 'center'
+    }
+  },
+  MuiListItem: {
+    root: {
+      paddingTop: 0,
+      paddingBottom: 0,
+      minHeight: '3.5rem'
+    },
+    gutters: {
+      paddingLeft: '1rem',
+      paddingRight: '1rem',
+      [theme.breakpoints.up('md')]: {
+        '&.listItem--dialog': {
+          paddingLeft: '2rem',
+          paddingRight: '2rem'
+        }
+      }
+    },
+    dense: {
+      minHeight: '3.5rem',
+      paddingTop: 0,
+      paddingBottom: 0
+    },
+    secondaryAction: {
+      paddingRight: '2rem'
+    },
+    button: {
+      '&$selected, &$selected:hover': {
+        backgroundColor: theme.palette.background.selected
+      },
+      '&:hover': {
+        backgroundColor: theme.palette.grey[100]
+      },
+      '&:focus': {
+        backgroundColor: theme.palette.grey[100]
+      }
+    }
+  },
+  MuiListItemText: {
+    root: {
+      padding: '14px 0'
+    },
+    dense: {
+      fontSize: null
+    },
+    primary: {
+      '&$dense': {
+        fontSize: null
+      }
+    },
+    secondary: {
+      '&$dense': {
+        fontSize: null
+      }
+    },
+    multiline: {
+      marginTop: 0,
+      marginBottom: 0
+    }
+  },
+  MuiListSubheader: {
+    root: {
+      borderTop: '1px solid transparent',
+      borderBottom: '1px solid transparent',
+      marginBottom: '-1px',
+      padding: 0,
+      height: '2rem',
+      backgroundColor: theme.palette.background.default,
+      textIndent: '2rem',
+      fontWeight: 'bold',
+      fontSize: '.75rem',
+      textTransform: 'uppercase',
+      alignItems: 'center',
+      display: 'flex',
+      lineHeight: 1.33,
+      color: theme.palette.text.secondary
+    },
+    gutters: {
+      paddingLeft: 0,
+      paddingRight: 0
+    },
+    sticky: {
+      backgroundColor: theme.palette.background.default
+    }
+  },
+  MuiListItemSecondaryAction: {
+    root: {
+      zIndex: 1,
+      right: 0
+    }
+  },
+  MuiMenuItem: {
+    root: {
+      minHeight: 'auto',
+      paddingTop: 4,
+      paddingBottom: 4,
+      color: theme.palette.grey[800],
+      border: 0,
+      '&:last-child': {
+        borderBottom: 0
+      }
+    },
+    gutters: {
+      paddingLeft: 24,
+      paddingRight: 24
+    }
+  },
+  MuiTextField: {
+    borderWidth: '0.0625rem'
+  },
+  MuiFormLabel: {
+    root: {
+      color: theme.palette.text.secondary
+    }
+  },
+  MuiFormHelperText: {
+    root: {
+      fontStyle: 'italic',
+      fontSize: '0.875rem',
+      marginTop: 4
+    }
+  },
+  MuiDialog: {
+    paper: {
+      '&.small': {
+        width: '480px',
+        maxWidth: '480px',
+        [theme.breakpoints.down('md')]: {
+          margin: '16px',
+          padding: '0 8px 8px',
+          height: 'auto',
+          maxHeight: 'calc(100% - 32px)',
+          borderRadius: '6px'
+        }
+      },
+      '&.medium': {
+        [theme.breakpoints.up('md')]: {
+          width: '544px',
+          maxWidth: '544px'
+        }
+      },
+      '&.large': {
+        [theme.breakpoints.up('md')]: {
+          width: '800px',
+          maxWidth: '800px'
+        }
+      }
+    },
+    scrollPaper: {
+      '&.alignTop': {
+        alignItems: 'start'
+      }
+    },
+    paperFullScreen: {
+      '& .cozyDialogActions': {
+        paddingBottom: 'env(safe-area-inset-bottom)'
+      }
+    }
+  },
+  MuiDialogTitle: {
+    root: {
+      ...theme.typography.h3,
+      width: 'calc(100% - (64px + 24px))', // remove padding and close button width + margin
+      padding: '24px 32px',
+      [theme.breakpoints.down('sm')]: {
+        ...theme.typography.h4,
+        width: 'calc(100% - 32px)', // remove padding
+        padding: '13px 16px 12px'
+      },
+      '&.dialogTitleFluid': {
+        width: '100%',
+        paddingTop: 0,
+        paddingLeft: 0,
+        paddingRight: 0,
+        [theme.breakpoints.down('sm')]: {
+          padding: '0 0 24px 0'
+        }
+      },
+      '&.dialogTitleFull': {
+        width: 'calc(100% - 64px)', // remove padding
+        [theme.breakpoints.down('sm')]: {
+          width: 'calc(100% - 32px)' // remove padding
+        }
+      }
+    }
+  },
+  MuiDialogContent: {
+    root: {
+      padding: '24px 32px 0',
+      [theme.breakpoints.down('sm')]: {
+        padding: '24px 16px 0'
+      },
+      '& .dialogContentInner': {
+        marginBottom: '24px',
+        '&.withFluidActions': {
+          [theme.breakpoints.down('sm')]: {
+            marginBottom: '16px'
+          }
+        }
+      }
+    }
+  },
+  MuiDialogActions: {
+    root: {
+      margin: '16px 32px',
+      padding: 0,
+      [theme.breakpoints.down('sm')]: {
+        margin: '8px 16px',
+        '& button': {
+          flexGrow: 1
+        }
+      },
+      '&.dialogActionsFluid': {
+        margin: '24px 0 0'
+      },
+      '&.columnLayout': {
+        display: 'flex',
+        flexDirection: 'column-reverse',
+        '& button': {
+          width: '100%',
+          margin: 0,
+          '&:not(:first-child)': {
+            marginBottom: '8px'
+          }
+        }
+      },
+
+      // To keep muiV3 behavior
+      // TODO check later if we need this behavior
+      '& > :not(:first-child):not(:first-child)': {
+        marginLeft: 4
+      }
+    }
+  },
+  MuiDivider: {
+    inset: {
+      marginLeft: 64,
+      '&.divider--dialog': {
+        marginLeft: 80
+      }
+    }
+  },
+  MuiSwitch: {
+    checked: {
+      '& + $track$track': {
+        opacity: 1
+      }
+    },
+    switchBase: {
+      top: 1,
+      '&$checked': {
+        transform: 'translateX(15px)'
+      }
+    },
+    thumb: {
+      width: 16,
+      height: 16
+    },
+    track: {
+      width: SWITCH_BAR_WIDTH,
+      height: 12,
+      backgroundColor: theme.palette.grey[200]
+    },
+    colorPrimary: {
+      '&$checked': {
+        color: getCssVariableValue('primaryContrastTextColor')
+      }
+    },
+    colorSecondary: {
+      '&$checked': {
+        '& + $track': {
+          backgroundColor: getCssVariableValue('successColor')
+        },
+        color: getCssVariableValue('primaryContrastTextColor')
+      }
+    },
+    disabled: {
+      '&$checked + $track': {
+        backgroundColor: `${theme.palette.grey[200]} !important`
+      },
+      '& $thumb': {
+        backgroundColor: 'white'
+      }
+    }
+  },
+  MuiTooltip: {
+    tooltip: {
+      // TODO should be grey[700] when grey[700] is supported
+      backgroundColor: theme.palette.grey[800],
+      borderRadius: '8px',
+      fontSize: '1rem',
+      color: 'white',
+      lineHeight: '1.3',
+      padding: '16px'
+    },
+    arrow: {
+      // TODO should be grey[700] when grey[700] is supported
+      color: theme.palette.grey[800]
+    },
+    popper: {
+      opacity: 0.9
+    }
+  },
+  MuiIconButton: {
+    root: {
+      color: theme.palette.text.secondary
+    }
+  }
+})
+
+export default makeOverrides
